feat: rewrite everything to use hyper directly
reqwest was giving me some headaches with regards to compatibility
with tower-http.
Instead of fighting with reqwest I decided to try working with hyper
directly, and so I started fighting with hyper instead. :)

Well, this was certainly a learning experience I suppose. The code is
kinda somewhat very janky. It honestly might be better to switch to the
legacy hyper client from hyper-util, but for now everything seems to
work okay.

If I run into issues later I can either do the switch then or try to
improve on my current implementation (and probably learn a bunch more
in the process!).

This commit also includes some random unrelated things, because during
the rewrite I had to touch various places of the codebase and I simply
couldn't be bothered to deal with the rewrite *and* maintain a neat
git history.

As far as I can remember the extra changes were:
- Switch to using tower-service and tower-layer directly for the
  implementors of those traits.
- A few bug fixes in the various API handlers.

Sorry for the big mess future me! :3

Signed-off-by: Linda Siemons <linda@catbrained.dev>