feat(client): add ability to reconnect
I moved the code for the connection establishment into a `Connector`
struct, so it can be reused when the connection gets closed due to
timeouts or other reasons.

Still not sure about the logic in the `InnerClient::send_request()`
function, but it works for now.
Also not 100% sure if there is a better way for sharing
a client/connection, but the wrapping in `Arc`s and cloning works
for now.

Signed-off-by: Linda Siemons <linda@catbrained.dev>